mobile phone service

+++ src/test/java/io/kirill/ebayapp/mobilephone/clients/cex/CexClientTest.java
@@ -52,7 +52,7 @@ class CexClientTest {
 
       StepVerifier
           .create(averagePrice)
-          .expectNextMatches(price -> price == 17.5)
+          .expectNextMatches(price -> price.doubleValue() == 17.5)
           .verifyComplete();
 
     var recordedRequest = mockWebServer.takeRequest();
